Create Celebrity Hobbies
In this project, I created two tables with information relating to celebrities. After creating the tables I wrote a query to join the hobbies table to the celebrities table on the celebrity name from each table to avoid duplicates. The resulting table produced the celebrity name, profession, hobby 1, and hobby 2 in one table.